add click handler definition in modebar button config:
- click handler are function of (modabar, eventObject)
