+++ src/components/modebar/index.js
@@ -13,7 +13,6 @@ var Plotly = require('../../plotly');
 var d3 = require('d3');
 
 var Icons = require('../../../build/ploticon');
-var buttonsConfig = require('./buttons_config');
 
 
 /**
@@ -237,284 +236,6 @@ proto.getLogo = function() {
 
     group.appendChild(a);
     return group;
-};
-
-/**
- * Apply D3 cartesian mode attributes to layout to update hover functionality
- * @Param {object} ev event object
- */
-proto.handleCartesian = function(ev) {
-    var button = ev.currentTarget,
-        astr = button.getAttribute('data-attr'),
-        val = button.getAttribute('data-val') || true,
-        _this = this,
-        graphInfo = this.graphInfo,
-        fullLayout = this.graphInfo._fullLayout,
-        aobj = {};
-
-    if(astr === 'zoom') {
-        var mag = (val === 'in') ? 0.5 : 2,
-            r0 = (1 + mag) / 2,
-            r1 = (1 - mag) / 2,
-            axList = Plotly.Axes.list(graphInfo, null, true);
-
-        var ax, axName, initialRange;
-
-        for(var i = 0; i < axList.length; i++) {
-            ax = axList[i];
-            if(!ax.fixedrange) {
-                axName = ax._name;
-                if(val === 'auto') aobj[axName + '.autorange'] = true;
-                else if(val === 'reset') {
-                    if(ax._rangeInitial === undefined) {
-                        aobj[axName + '.autorange'] = true;
-                    }
-                    else aobj[axName + '.range'] = ax._rangeInitial.slice();
-                }
-                else {
-                    initialRange = ax.range;
-                    aobj[axName + '.range'] = [
-                        r0 * initialRange[0] + r1 * initialRange[1],
-                        r0 * initialRange[1] + r1 * initialRange[0]
-                    ];
-                }
-            }
-        }
-    } else {
-        // if ALL traces have orientation 'h', 'hovermode': 'x' otherwise: 'y'
-        if (astr==='hovermode' && (val==='x' || val==='y')) {
-            val = fullLayout._isHoriz ? 'y' : 'x';
-            button.setAttribute('data-val', val);
-        }
-
-        aobj[astr] = val;
-    }
-
-    Plotly.relayout(graphInfo, aobj).then( function() {
-        _this.updateActiveButton();
-        if(astr === 'dragmode') {
-            if(fullLayout._hasCartesian) {
-                Plotly.Fx.setCursor(
-                    fullLayout._paper.select('.nsewdrag'),
-                    {pan:'move', zoom:'crosshair'}[val]
-                );
-            }
-            Plotly.Fx.supplyLayoutDefaults(graphInfo.layout, fullLayout,
-                graphInfo._fullData);
-        }
-    });
-};
-
-/**
- * Toggle the data hover mode
- * @Param {object} ev event object
- */
-proto.handleHover3d = function(ev) {
-    var button = ev.currentTarget,
-        val = JSON.parse(button.getAttribute('data-val')) || false,
-        _this = this,
-        graphInfo = this.graphInfo,
-        fullLayout = graphInfo._fullLayout,
-        sceneIds = Plotly.Plots.getSubplotIds(fullLayout, 'gl3d'),
-        layoutUpdate = {},
-
-        // initialize 'current spike' object to be stored in the DOM
-        currentSpikes = {},
-        axes = ['xaxis', 'yaxis', 'zaxis'],
-        spikeAttrs = ['showspikes', 'spikesides', 'spikethickness', 'spikecolor'];
-
-    var i, sceneId, sceneLayout, sceneSpikes;
-    var j, axis, axisSpikes;
-    var k, spikeAttr;
-
-    if (val) {
-        layoutUpdate = val;
-        button.setAttribute('data-val', JSON.stringify(null));
-    }
-    else {
-        layoutUpdate = {'allaxes.showspikes': false};
-
-        for (i = 0;  i < sceneIds.length; i++) {
-            sceneId = sceneIds[i];
-            sceneLayout = fullLayout[sceneId];
-            sceneSpikes = currentSpikes[sceneId] = {};
-
-            // copy all the current spike attrs
-            for (j = 0; j < 3; j++) {
-                axis = axes[j];
-                axisSpikes = sceneSpikes[axis] = {};
-                for (k = 0; k < spikeAttrs.length; k++) {
-                    spikeAttr = spikeAttrs[k];
-                    axisSpikes[spikeAttr] = sceneLayout[axis][spikeAttr];
-                }
-            }
-        }
-
-        button.setAttribute('data-val', JSON.stringify(currentSpikes));
-    }
-
-    Plotly.relayout(graphInfo, layoutUpdate).then( function() {
-        _this.updateActiveButton(button);
-    });
-
-};
-
-/**
- * Reconfigure keyboard bindings for webgl3D camera control on drag
- * @Param {object} ev event object
- */
-proto.handleDrag3d = function(ev) {
-    var button = ev.currentTarget,
-        attr = button.getAttribute('data-attr'),
-        val = button.getAttribute('data-val') || true,
-        _this = this,
-        graphInfo = this.graphInfo,
-        layoutUpdate = {};
-
-    layoutUpdate[attr] = val;
-
-    // Dragmode will go through the relayout->doplot->scene.plot()
-    // routine where the dragmode will be set in scene.plot()
-    Plotly.relayout(graphInfo, layoutUpdate).then( function() {
-        _this.updateActiveButton();
-    });
-};
-
-
-/**
- * Reset the position of the webgl3D camera
- * @Param {object} ev event object
- */
-proto.handleCamera3d = function(ev) {
-    var button = ev.currentTarget,
-        attr = button.getAttribute('data-attr'),
-        layout = this.graphInfo.layout,
-        fullLayout = this.graphInfo._fullLayout,
-        sceneIds = Plotly.Plots.getSubplotIds(fullLayout, 'gl3d');
-
-    var i, sceneId, sceneLayout, fullSceneLayout, scene, cameraPos;
-
-    for (i = 0;  i < sceneIds.length; i++) {
-        sceneId = sceneIds[i];
-        sceneLayout = layout[sceneId];
-        fullSceneLayout = fullLayout[sceneId];
-        scene = fullSceneLayout._scene;
-
-        if (!sceneLayout || attr==='resetDefault') scene.setCameraToDefault();
-        else if (attr === 'resetLastSave') {
-
-            cameraPos = sceneLayout.camera;
-            if (cameraPos) scene.setCamera(cameraPos);
-            else scene.setCameraToDefault();
-        }
-    }
-
-    /* TODO have a sceneLastTouched in _fullLayout to only
-     * update the camera of the scene last touched by the user
-     */
-};
-
-proto.handleGeo = function(ev) {
-    var button = ev.currentTarget,
-        attr = button.getAttribute('data-attr'),
-        val = button.getAttribute('data-val') || true,
-        fullLayout = this.graphInfo._fullLayout,
-        geoIds = Plotly.Plots.getSubplotIds(fullLayout, 'geo');
-
-    var geo, scale, newScale;
-
-    for(var i = 0;  i < geoIds.length; i++) {
-        geo = fullLayout[geoIds[i]]._geo;
-
-        if(attr === 'zoom') {
-            scale = geo.projection.scale();
-            newScale = val==='in' ? 2 * scale : 0.5 * scale;
-            geo.projection.scale(newScale);
-            geo.zoom.scale(newScale);
-            geo.render();
-        }
-        else if(attr === 'reset') geo.zoomReset();
-        else if(attr === 'hovermode') geo.showHover = !geo.showHover;
-    }
-
-    this.updateActiveButton(button);
-};
-
-proto.handleHoverPie = function() {
-    var _this = this,
-        graphInfo = _this.graphInfo,
-        newHover = graphInfo._fullLayout.hovermode ?
-            false :
-            'closest';
-
-    Plotly.relayout(graphInfo, 'hovermode', newHover).then(function() {
-        _this.updateActiveButton();
-    });
-};
-
-proto.handleHoverGl2d = function(ev) {
-    var _this = this,
-        button  = ev.currentTarget,
-        graphInfo = _this.graphInfo,
-        newHover = graphInfo._fullLayout.hovermode ?  false : 'closest';
-
-    Plotly.relayout(graphInfo, 'hovermode', newHover).then(function() {
-        _this.updateActiveButton(button);
-    });
-};
-
-proto.toImage = function() {
-
-    var format = 'png';
-    var _this = this;
-
-    if ( Plotly.Lib.isIE() ) {
-        Plotly.Lib.notifier('Snapshotting is unavailable in Internet Explorer. ' +
-                            'Consider exporting your images using the Plotly Cloud', 'long');
-        return;
-    }
-
-    if (this._snapshotInProgress) {
-        Plotly.Lib.notifier('Snapshotting is still in progress - please hold', 'long');
-        return;
-    }
-
-    this._snapshotInProgress = true;
-    Plotly.Lib.notifier('Taking snapshot - this may take a few seconds', 'long');
-
-    var ev = Plotly.Snapshot.toImage(this.graphInfo, {format: format});
-
-    var filename = this.graphInfo.fn || 'newplot';
-    filename += '.' + format;
-
-    ev.once('success', function(result) {
-
-        _this._snapshotInProgress = false;
-
-        var downloadLink = document.createElement('a');
-        downloadLink.href = result;
-        downloadLink.download = filename; // only supported by FF and Chrome
-
-        document.body.appendChild(downloadLink);
-        downloadLink.click();
-        document.body.removeChild(downloadLink);
-
-        ev.clean();
-    });
-
-    ev.once('error', function (err) {
-        _this._snapshotInProgress = false;
-
-        Plotly.Lib.notifier('Sorry there was a problem downloading your ' + format, 'long');
-        console.error(err);
-
-        ev.clean();
-    });
-};
-
-proto.sendDataToCloud = function() {
-    var gd = this.graphInfo;
-    Plotly.Plots.sendDataToCloud(gd);
 };
 
 proto.removeAllButtons = function() {
